refactoring

+++ src/main/kotlin/gq/learningEnglish/model/questionnaire/Question.kt
@@ -7,7 +7,7 @@ data class Question(
     @DbField(value = "asking_language")
     val askingLanguage: AvailableLanguages,
     @DbField(value = "asking_word_id")
-    val askingWordId: Int,
+    val askingWordId: Long,
     @DbField(value = "asking_word")
     val askingWord: String,
     @DbField
